Initial commit
The README and COPYING files are included to set up the Mafia Game Suite
repo. With the instructions in the README, the Suite can be used for a
full game, albeit not fully automatically.

An empty `game` file is included to direct the users' attention to it,
since it's the most important file in the Suite.

`generate_roles.sh` is the first script to be created for the Suite. It
relies on the `players` and `roles` files described in the README, to
produce the `secret` file containing a shuffled list of players and
roles ready for play.