Add kill.sh script
kill.sh makes use of two files: dead and log. dead is a record of which
players have been killed, while log is a record of all actions.

kill.sh --nop allows appending a "kill" action to the log but without
actually killing the player, e.g. the action was blocked by a Medic.
This guarantees that dead will only contain dead players, and log
contains the full transcription of actions regardless of success or
failure.